[test] apply the new test style to databend-query - interpreters (81/177)
