+++ query/tests/it/interpreters/interpreter_database_create.rs
@@ -15,22 +15,21 @@
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::interpreters::*;
 use futures::stream::StreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::interpreters::*;
-use crate::sql::*;
+use crate::tests::parse_query;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_create_user_interpreter() -> Result<()> {
+async fn test_create_database_interpreter() -> Result<()> {
     common_tracing::init_default_ut_tracing();
 
     let ctx = crate::tests::create_query_context()?;
 
-    static TEST_QUERY: &str = "CREATE USER 'test'@'localhost' IDENTIFIED BY 'password'";
-    if let PlanNode::CreateUser(plan) = PlanParser::parse(TEST_QUERY, ctx.clone()).await? {
-        let executor = CreatUserInterpreter::try_create(ctx, plan.clone())?;
-        assert_eq!(executor.name(), "CreateUserInterpreter");
+    if let PlanNode::CreateDatabase(plan) = parse_query("create database db1", &ctx)? {
+        let executor = CreateDatabaseInterpreter::try_create(ctx, plan.clone())?;
+        assert_eq!(executor.name(), "CreateDatabaseInterpreter");
         let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
     } else {
