+++ query/tests/it/interpreters/interpreter_database_drop.rs
@@ -15,23 +15,23 @@
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
-use futures::stream::StreamExt;
+use databend_query::interpreters::*;
+use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::interpreters::*;
 use crate::tests::parse_query;
 
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_create_database_interpreter() -> Result<()> {
-    common_tracing::init_default_ut_tracing();
-
+#[tokio::test]
+async fn test_drop_database_interpreter() -> Result<()> {
     let ctx = crate::tests::create_query_context()?;
 
-    if let PlanNode::CreateDatabase(plan) = parse_query("create database db1", &ctx)? {
-        let executor = CreateDatabaseInterpreter::try_create(ctx, plan.clone())?;
-        assert_eq!(executor.name(), "CreateDatabaseInterpreter");
-        let mut stream = executor.execute(None).await?;
-        while let Some(_block) = stream.next().await {}
+    if let PlanNode::DropDatabase(plan) = parse_query("drop database default", &ctx)? {
+        let executor = DropDatabaseInterpreter::try_create(ctx, plan.clone())?;
+        assert_eq!(executor.name(), "DropDatabaseInterpreter");
+        let stream = executor.execute(None).await?;
+        let result = stream.try_collect::<Vec<_>>().await?;
+        let expected = vec!["++", "++"];
+        common_datablocks::assert_blocks_sorted_eq(expected, result.as_slice());
     } else {
         panic!()
     }
