+++ query/tests/it/interpreters/interpreter_setting.rs
@@ -15,21 +15,21 @@
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::interpreters::*;
 use futures::stream::StreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::interpreters::*;
 use crate::tests::parse_query;
 
-#[tokio::test]
-async fn test_use_interpreter() -> Result<()> {
+#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+async fn test_setting_interpreter() -> Result<()> {
     let ctx = crate::tests::create_query_context()?;
 
-    if let PlanNode::UseDatabase(plan) = parse_query("USE default", &ctx)? {
-        let interpreter = UseDatabaseInterpreter::try_create(ctx, plan)?;
-        assert_eq!(interpreter.name(), "UseDatabaseInterpreter");
+    if let PlanNode::SetVariable(plan) = parse_query("SET max_block_size=1", &ctx)? {
+        let executor = SettingInterpreter::try_create(ctx, plan)?;
+        assert_eq!(executor.name(), "SettingInterpreter");
 
-        let mut stream = interpreter.execute(None).await?;
+        let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
     } else {
         panic!()
@@ -38,15 +38,14 @@ async fn test_use_interpreter() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
-async fn test_use_database_interpreter_error() -> Result<()> {
+#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+async fn test_setting_interpreter_error() -> Result<()> {
     let ctx = crate::tests::create_query_context()?;
 
-    if let PlanNode::UseDatabase(plan) = parse_query("USE xx", &ctx)? {
-        let interpreter = UseDatabaseInterpreter::try_create(ctx, plan)?;
-
-        if let Err(e) = interpreter.execute(None).await {
-            let expect = "Code: 3, displayText = Cannot USE 'xx', because the 'xx' doesn't exist.";
+    if let PlanNode::SetVariable(plan) = parse_query("SET xx = 1", &ctx)? {
+        let executor = SettingInterpreter::try_create(ctx, plan)?;
+        if let Err(e) = executor.execute(None).await {
+            let expect = "Code: 20, displayText = Unknown variable: \"xx\".";
             assert_eq!(expect, format!("{}", e));
         } else {
             panic!();
